\name{FinePop2-package}
\alias{FinePop2-package}
\docType{package}
\title{\packageTitle{FinePop2}}
\description{\packageDescription{FinePop2}}
\author{
\packageAuthor{FinePop2}\cr
Maintainer: \packageMaintainer{FinePop2}
}
\references{
Kitada S, Nakamichi R, Kishino H (2017) The empirical Bayes estimators of fine-scale population structure in high gene flow species. \emph{Mol. Ecol. Resources}, 17, 1210-1222.\cr
Nei M, Chesser RK (1983) Estimation of fixation indices and gene diversity. \emph{Annals of Human Genetics}, 47, 253-259.\cr
Rousset F (2008) Genepop'007: a complete reimplementation of the Genepop software for Windows and Linux. \emph{Mol. Ecol. Resources}, 8, 103-106.\cr
Weir BS, Cockerham CC (1984) Estimating F-statistics for the analysis of population structure. \emph{Evolution}, 38, 1358-1370.\cr
Weir BS, Goudet J (2017) A Unified Characterization of Population Structure and Relatedness. \emph{Genetics}, 206, 2085-2103.
}
\keyword{ package }
